separated arm and intake subsystems and added interlocking command conditions
