+++ src/main/java/frc/robot/RobotContainer.java
@@ -9,19 +9,28 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.RunCommand;
+import edu.wpi.first.wpilibj2.command.StartEndCommand;
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
 import frc.robot.Constants.OIConstants;
 import frc.robot.arm.Arm;
 import frc.robot.drivetrain.Drivetrain;
 import frc.robot.drivetrain.commands.ZorroDrive;
+import frc.robot.intake.Intake;
 
 public class RobotContainer {
 
   private final Drivetrain m_swerve = new Drivetrain();
   private final Arm m_arm = new Arm();
+  private final Intake m_intake = new Intake();
 
   private Joystick m_driver = new Joystick(OIConstants.kDriverControllerPort);
   private XboxController m_operator = new XboxController(OIConstants.kOperatorControllerPort);
+
+  private StartEndCommand lowerArmCommand =
+      new StartEndCommand(() -> m_arm.pneumaticRetract(), () -> {}, m_arm);
+
+  private StartEndCommand raiseArmCommand =
+      new StartEndCommand(() -> m_arm.pneumaticDeploy(), () -> {}, m_arm);
 
   public RobotContainer() {
 
@@ -37,17 +46,26 @@ public class RobotContainer {
         .onTrue(new InstantCommand(() -> m_swerve.resetGyro()).ignoringDisable(true));
 
     // Operator controller buttons
-    new JoystickButton(m_operator, Button.kLeftBumper.value)
-        .onTrue(new InstantCommand(() -> m_arm.pneumaticRetract()));
+    new JoystickButton(m_operator, Button.kLeftBumper.value).onTrue(lowerArmCommand);
+    new JoystickButton(m_operator, Button.kRightBumper.value).onTrue(raiseArmCommand);
 
-    new JoystickButton(m_operator, Button.kRightBumper.value)
-        .onTrue(new InstantCommand(() -> m_arm.pneumaticDeploy()));
-
+    // Intake Note from floor
     new JoystickButton(m_operator, Button.kX.value)
-        .whileTrue(new RunCommand(() -> m_arm.setVelocity(1.0)).until(m_arm.m_noteSensor::get));
+        .whileTrue(
+            new RunCommand(() -> m_intake.setVelocity(1.0))
+                .until(m_intake.m_noteSensor::get)
+                .onlyIf(lowerArmCommand::isScheduled));
 
+    // Shift Note further into Intake
+    new JoystickButton(m_operator, Button.kA.value)
+        .onTrue(
+            new InstantCommand(() -> m_intake.resetIntakeEncoder())
+                .andThen(new RunCommand(() -> m_intake.setPosition(0.2))));
+
+    // Shoot Note into Amp
     new JoystickButton(m_operator, Button.kY.value)
-        .whileTrue(new RunCommand(() -> m_arm.setVelocity(1.0)));
+        .whileTrue(
+            new RunCommand(() -> m_intake.setVelocity(1.0)).onlyIf(raiseArmCommand::isScheduled));
   }
 
   public Command getAutonomousCommand() {
